+++ src/components/Topic/Full.tsx
@@ -184,17 +184,6 @@ export const FullTopic = (props: Props) => {
     }
   }
 
-  // 🔧 ОТЛАДКА: Проверим props топика
-  console.log('[FullTopic] Props received:', {
-    topicTitle: props.topic?.title,
-    hasTopic: !!props.topic,
-    hasTopicStat: !!props.topic?.stat,
-    topicStat: props.topic?.stat,
-    shoutsCount: props.topic?.stat?.shouts,
-    authorsCount: props.topic?.stat?.authors,
-    followersCount: props.topic?.stat?.followers
-  })
-
   return (
     <div class={clsx(styles.topicHeader, 'col-md-16 col-lg-12 offset-md-4 offset-lg-6')}>
       <h1>{title()}</h1>